cmpe273-fall20-master/requirements.txt
pyzmq == 19.0.2
mmh3
python-consul